+++ R/shinyscholar-package.R
@@ -14,5 +14,4 @@
 #'   function \code{\link{run_shinyscholar}} and requires
 #'   \code{install.packages("shinyscholar", dependencies = TRUE)}.
 #'
-#' @import bslib gargoyle leaflet shiny
 NULL