+++ tests/testthat/test-create_template.R
@@ -129,6 +129,11 @@ test_that("Check create template returns expected errors", {
 })
 
 test_that("Check create template function works as expected", {
+
+  withr::local_envvar(
+    R_USER_CACHE_DIR = tempfile(),
+  )
+
   withr::with_temp_libpaths({
     modules$map <- c(TRUE, TRUE, FALSE, FALSE)
 
@@ -153,7 +158,7 @@ test_that("Check create template function works as expected", {
 
     if (suggests){
       devtools::document(file.path(directory, name))
-      devtools::install(file.path(directory, name), force = TRUE, quick = TRUE, dependencies = FALSE)
+      pak::local_install(file.path(directory, name), dependencies = FALSE)
 
       app <- shinytest2::AppDriver$new(app_dir = file.path(directory, name, "inst", "shiny"), name = "create_test")
       common <- app$get_value(export = "common")
@@ -164,6 +169,11 @@ test_that("Check create template function works as expected", {
 })
 
 test_that("Check create template function works with false settings", {
+
+  withr::local_envvar(
+    R_USER_CACHE_DIR = tempfile(),
+  )
+
   withr::with_temp_libpaths({
     modules$map <- c(FALSE, FALSE, FALSE, FALSE)
     modules$result <- c(TRUE, FALSE, FALSE, FALSE)
@@ -195,7 +205,7 @@ test_that("Check create template function works with false settings", {
 
     if (suggests){
       devtools::document(file.path(directory, name))
-      devtools::install(file.path(directory, name), force = TRUE, quick = TRUE, dependencies = FALSE)
+      pak::local_install(file.path(directory, name), dependencies = FALSE)
 
       app <- shinytest2::AppDriver$new(app_dir = file.path(directory, name, "inst", "shiny"), name = "create_test")
       common <- app$get_value(export = "common")
@@ -207,6 +217,11 @@ test_that("Check create template function works with false settings", {
 
 test_that("Check async, no map runs correctly", {
   if (suggests){
+
+    withr::local_envvar(
+      R_USER_CACHE_DIR = tempfile(),
+    )
+
     withr::with_temp_libpaths({
       modules$map <- c(FALSE, FALSE, FALSE, FALSE)
       modules$result <- c(TRUE, FALSE, FALSE, FALSE)
@@ -225,7 +240,7 @@ test_that("Check async, no map runs correctly", {
                   include_table = FALSE, include_code = FALSE, install = FALSE)
 
       devtools::document(file.path(directory, name))
-      devtools::install(file.path(directory, name), force = TRUE, quick = TRUE, dependencies = FALSE)
+      pak::local_install(file.path(directory, name), dependencies = FALSE)
 
       app <- shinytest2::AppDriver$new(app_dir = file.path(directory, name, "inst", "shiny"), name = "create_test")
       common <- app$get_value(export = "common")
@@ -237,6 +252,11 @@ test_that("Check async, no map runs correctly", {
 
 test_that("Check async, with map runs correctly", {
   if (suggests){
+
+    withr::local_envvar(
+      R_USER_CACHE_DIR = tempfile(),
+    )
+
     withr::with_temp_libpaths({
       modules$map <- c(TRUE, FALSE, FALSE, FALSE)
       modules$result <- c(TRUE, FALSE, FALSE, FALSE)
@@ -255,7 +275,7 @@ test_that("Check async, with map runs correctly", {
                       include_table = FALSE, include_code = TRUE, install = FALSE)
 
       devtools::document(file.path(directory, name))
-      devtools::install(file.path(directory, name), force = TRUE, quick = TRUE, dependencies = FALSE)
+      pak::local_install(file.path(directory, name), dependencies = FALSE)
 
       app <- shinytest2::AppDriver$new(app_dir = file.path(directory, name, "inst", "shiny"), name = "create_test")
       common <- app$get_value(export = "common")