+++ tests/testthat/test-save_load.R
@@ -1,13 +1,4 @@
 if (suggests){
-  test_that("{shinytest2} recording: e2e_empty_save", {
-
-    skip_if(is_fedora())
-    skip_on_cran()
-
-    rerun_test("empty_save_test", list(save_path = save_path))
-    common <- readRDS(save_path)
-    expect_true(is.null(common$raster))
-  })
 
   test_that("{shinytest2} recording: e2e_save_scat", {
 
@@ -64,10 +55,15 @@ if (suggests){
       )
     )
 
-    common <- common_class$new()
-    common$state$main$version <- as.character(packageVersion("shinyscholar"))
-    common$state$main$app <- "shinyscholar"
-    saveRDS(common, save_path)
+    temp <- as.list(common_class$new())
+    common_items <- names(temp)
+    # exclude the non-public, function objects and tasks
+    save_items  <- common_items[!common_items %in% c("clone", ".__enclos_env__")]
+    temp <- temp[save_items]
+    temp$state$main$version <- as.character(packageVersion("shinyscholar"))
+    temp$state$main$app <- "shinyscholar"
+    class(temp) <- "common"
+    saveRDS(temp, save_path)
 
     app <- shinytest2::AppDriver$new(app_dir = system.file("shiny", package = "shinyscholar"), name = "e2e_load")
     app$set_inputs(introTabs = "Load Prior Session")