+++ NEWS.md
@@ -136,3 +136,5 @@ shinyscholar 0.4.5
 - Migrate from `{devtools}` to `{pak}` for installation
 - Set `R_USER_CACHE_DIR` in tests
 - Import app dependencies in `global.R` instead of via `Depends`
+- Fixed bug in `global.R` due to missing function
+- `create_module()` updates the module `.yml` and `global.R` so that the app runs