+++ NEWS.md
@@ -138,3 +138,5 @@ shinyscholar 0.4.5
 - Import app dependencies in `global.R` instead of via `Depends`
 - Fixed bug in `global.R` due to missing function
 - `create_module()` updates the module `.yml` and `global.R` so that the app runs
+- `reset_data()`, `show_loading_modal()`, `close_loading_modal()`, `show_map()`, `show_results()` and `show_table()` have been moved from `R/helper_functions` to `inst/shiny/ui_helpers.R` and are no longer exported
+