+++ NEWS.md
@@ -119,3 +119,24 @@ shinyscholar 0.4.2
 - Various tweaks to ensure that created apps pass `R CMD check`
 - Removed `{magrittr}` dependency and migrated to native pipes, increasing R dependency to v4.1.0
 - `create_module()` now creates skeleton functions that include `{roxygen2}` tags
+
+shinyscholar 0.4.3
+=============
+- `create_template()` now creates `testthat.R` file
+- Fixed `create_module()` module function calls for async apps containing a map
+- Saving apps now only stores data in `common` (i.e. excluding `common$tasks`)
+- Updated README with clarifications
+
+shinyscholar 0.4.4
+=============
+- Skip problematic test on CRAN
+
+shinyscholar 0.4.5
+=============
+- Migrate from `{devtools}` to `{pak}` for installation
+- Set `R_USER_CACHE_DIR` in tests
+- Import app dependencies in `global.R` instead of via `Depends`
+- Fixed bug in `global.R` due to missing function
+- `create_module()` updates the module `.yml` and `global.R` so that the app runs
+- `reset_data()`, `show_loading_modal()`, `close_loading_modal()`, `show_map()`, `show_results()` and `show_table()` have been moved from `R/helper_functions` to `inst/shiny/ui_helpers.R` and are no longer exported
+